move shadcn to app
