+++ app/ui/carousel.tsx
@@ -5,7 +5,7 @@ import useEmblaCarousel, {type UseEmblaCarouselType} from 'embla-carousel-react'
 import {ArrowLeft, ArrowRight} from 'lucide-react'
 
 import {cn} from '@/lib/utils'
-import {Button} from '@/components/ui/button'
+import {Button} from '@/app/ui/button'
 
 type CarouselApi = UseEmblaCarouselType[1]
 type UseCarouselParameters = Parameters<typeof useEmblaCarousel>
